Server Serialization & other improvements
Created a Directory object to be serialized and move all data to that
object.
Changing parameters to be more consistent - using “auth” as the
requester’s username.
Added new /isValidUser endpoint for checking if a username exists. This
differs from /checkUser because that overwrites the username’s ip with
the requester’s ip.